@@ -1,9 +1,14 @@
-﻿using MessagePack.Formatters;
-using Reactive.Bindings;
+﻿// Copyright (c) All contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
 using System;
 using System.Collections.Generic;
 using System.Reactive;
 using System.Reactive.Concurrency;
+using MessagePack.Formatters;
+using Reactive.Bindings;
+
+#pragma warning disable SA1649 // File name should match first type name
 
 namespace MessagePack.ReactivePropertyExtension
 {
@@ -11,25 +16,25 @@ namespace MessagePack.ReactivePropertyExtension
     {
 #pragma warning disable CS0618
 
-        static Dictionary<int, IScheduler> mapTo = new Dictionary<int, IScheduler>();
-        static Dictionary<IScheduler, int> mapFrom = new Dictionary<IScheduler, int>();
+        private static Dictionary<int, IScheduler> mapTo = new Dictionary<int, IScheduler>();
+        private static Dictionary<IScheduler, int> mapFrom = new Dictionary<IScheduler, int>();
 
         static ReactivePropertySchedulerMapper()
         {
             // default map
-            var mappings = new[]
+            (int, IScheduler)[] mappings = new[]
             {
-               (-2, CurrentThreadScheduler.Instance ),
-               (-3, ImmediateScheduler.Instance ),
-               (-4, TaskPoolScheduler.Default ),
-               (-5, System.Reactive.Concurrency.NewThreadScheduler.Default ),
-               (-6, Scheduler.ThreadPool ),
+               (-2, CurrentThreadScheduler.Instance),
+               (-3, ImmediateScheduler.Instance),
+               (-4, TaskPoolScheduler.Default),
+               (-5, System.Reactive.Concurrency.NewThreadScheduler.Default),
+               (-6, Scheduler.ThreadPool),
                (-7, System.Reactive.Concurrency.DefaultScheduler.Instance),
 
-               (-1, UIDispatcherScheduler.Default ), // override
+               (-1, UIDispatcherScheduler.Default), // override
             };
 
-            foreach (var item in mappings)
+            foreach ((int, IScheduler) item in mappings)
             {
                 ReactivePropertySchedulerMapper.mapTo[item.Item1] = item.Item2;
                 ReactivePropertySchedulerMapper.mapFrom[item.Item2] = item.Item1;
@@ -72,11 +77,12 @@ namespace MessagePack.ReactivePropertyExtension
 
         internal static int ToReactivePropertyModeInt<T>(global::Reactive.Bindings.ReactiveProperty<T> reactiveProperty)
         {
-            var mode = ReactivePropertyMode.None;
+            ReactivePropertyMode mode = ReactivePropertyMode.None;
             if (reactiveProperty.IsDistinctUntilChanged)
             {
                 mode |= ReactivePropertyMode.DistinctUntilChanged;
             }
+
             if (reactiveProperty.IsRaiseLatestValueOnSubscribe)
             {
                 mode |= ReactivePropertyMode.RaiseLatestValueOnSubscribe;
@@ -87,15 +93,17 @@ namespace MessagePack.ReactivePropertyExtension
 
         internal static int ToReactivePropertySlimModeInt<T>(global::Reactive.Bindings.ReactivePropertySlim<T> reactiveProperty)
         {
-            var mode = ReactivePropertyMode.None;
+            ReactivePropertyMode mode = ReactivePropertyMode.None;
             if (reactiveProperty.IsDistinctUntilChanged)
             {
                 mode |= ReactivePropertyMode.DistinctUntilChanged;
             }
+
             if (reactiveProperty.IsRaiseLatestValueOnSubscribe)
             {
                 mode |= ReactivePropertyMode.RaiseLatestValueOnSubscribe;
             }
+
             return (int)mode;
         }
     }
@@ -103,79 +111,78 @@ namespace MessagePack.ReactivePropertyExtension
     // [Mode, SchedulerId, Value] : length should be three.
     public class ReactivePropertyFormatter<T> : IMessagePackFormatter<ReactiveProperty<T>>
     {
-        public int Serialize(ref byte[] bytes, int offset, ReactiveProperty<T> value, IFormatterResolver formatterResolver)
+        public void Serialize(ref MessagePackWriter writer, ReactiveProperty<T> value, MessagePackSerializerOptions options)
         {
             if (value == null)
             {
-                return MessagePackBinary.WriteNil(ref bytes, offset);
+                writer.WriteNil();
             }
             else
             {
-                var startOffset = offset;
+                writer.WriteArrayHeader(3);
 
-                offset += MessagePackBinary.WriteFixedArrayHeaderUnsafe(ref bytes, offset, 3);
-
-                offset += MessagePackBinary.WriteInt32(ref bytes, offset, ReactivePropertySchedulerMapper.ToReactivePropertyModeInt(value));
-                offset += MessagePackBinary.WriteInt32(ref bytes, offset, ReactivePropertySchedulerMapper.GetSchedulerId(value.RaiseEventScheduler));
-                offset += formatterResolver.GetFormatterWithVerify<T>().Serialize(ref bytes, offset, value.Value, formatterResolver);
-
-                return offset - startOffset;
+                writer.Write(ReactivePropertySchedulerMapper.ToReactivePropertyModeInt(value));
+                writer.Write(ReactivePropertySchedulerMapper.GetSchedulerId(value.RaiseEventScheduler));
+                options.Resolver.GetFormatterWithVerify<T>().Serialize(ref writer, value.Value, options);
             }
         }
 
-        public ReactiveProperty<T> Deserialize(byte[] bytes, int offset, IFormatterResolver formatterResolver, out int readSize)
+        public ReactiveProperty<T> Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
         {
-            if (MessagePackBinary.IsNil(bytes, offset))
+            if (reader.TryReadNil())
             {
-                readSize = 1;
                 return null;
             }
             else
             {
-                var startOffset = offset;
+                var length = reader.ReadArrayHeader();
+                if (length != 3)
+                {
+                    throw new InvalidOperationException("Invalid ReactiveProperty data.");
+                }
 
-                var length = MessagePackBinary.ReadArrayHeader(bytes, offset, out readSize);
-                offset += readSize;
-                if (length != 3) throw new InvalidOperationException("Invalid ReactiveProperty data.");
+                var mode = (ReactivePropertyMode)reader.ReadInt32();
 
-                var mode = (ReactivePropertyMode)MessagePackBinary.ReadInt32(bytes, offset, out readSize);
-                offset += readSize;
+                var schedulerId = reader.ReadInt32();
 
-                var schedulerId = MessagePackBinary.ReadInt32(bytes, offset, out readSize);
-                offset += readSize;
+                IScheduler scheduler = ReactivePropertySchedulerMapper.GetScheduler(schedulerId);
 
-                var scheduler = ReactivePropertySchedulerMapper.GetScheduler(schedulerId);
+                options.Security.DepthStep(ref reader);
+                try
+                {
+                    T v = options.Resolver.GetFormatterWithVerify<T>().Deserialize(ref reader, options);
 
-                var v = formatterResolver.GetFormatterWithVerify<T>().Deserialize(bytes, offset, formatterResolver, out readSize);
-                offset += readSize;
-
-                readSize = offset - startOffset;
-
-                return new ReactiveProperty<T>(scheduler, v, mode);
+                    return new ReactiveProperty<T>(scheduler, v, mode);
+                }
+                finally
+                {
+                    reader.Depth--;
+                }
             }
-
         }
     }
 
     public class InterfaceReactivePropertyFormatter<T> : IMessagePackFormatter<IReactiveProperty<T>>
     {
-        public int Serialize(ref byte[] bytes, int offset, IReactiveProperty<T> value, IFormatterResolver formatterResolver)
+        public void Serialize(ref MessagePackWriter writer, IReactiveProperty<T> value, MessagePackSerializerOptions options)
         {
             var rxProp = value as ReactiveProperty<T>;
             if (rxProp != null)
             {
-                return ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactiveProperty<T>>().Serialize(ref bytes, offset, rxProp, formatterResolver);
+                ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactiveProperty<T>>().Serialize(ref writer, rxProp, options);
+                return;
             }
 
             var slimProp = value as ReactivePropertySlim<T>;
             if (slimProp != null)
             {
-                return ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactivePropertySlim<T>>().Serialize(ref bytes, offset, slimProp, formatterResolver);
+                ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactivePropertySlim<T>>().Serialize(ref writer, slimProp, options);
+                return;
             }
 
             if (value == null)
             {
-                return MessagePackBinary.WriteNil(ref bytes, offset);
+                writer.WriteNil();
             }
             else
             {
@@ -183,41 +190,51 @@ namespace MessagePack.ReactivePropertyExtension
             }
         }
 
-        public IReactiveProperty<T> Deserialize(byte[] bytes, int offset, IFormatterResolver formatterResolver, out int readSize)
+        public IReactiveProperty<T> Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
         {
-            var length = MessagePackBinary.ReadArrayHeader(bytes, offset, out readSize);
+            var length = reader.ReadArrayHeader();
 
-            switch (length)
+            options.Security.DepthStep(ref reader);
+            try
             {
-                case 2:
-                    return ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactivePropertySlim<T>>().Deserialize(bytes, offset, formatterResolver, out readSize);
-                case 3:
-                    return ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactiveProperty<T>>().Deserialize(bytes, offset, formatterResolver, out readSize);
-                default:
-                    throw new InvalidOperationException("Invalid ReactiveProperty or ReactivePropertySlim data.");
+                switch (length)
+                {
+                    case 2:
+                        return ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactivePropertySlim<T>>().Deserialize(ref reader, options);
+                    case 3:
+                        return ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactiveProperty<T>>().Deserialize(ref reader, options);
+                    default:
+                        throw new InvalidOperationException("Invalid ReactiveProperty or ReactivePropertySlim data.");
+                }
+            }
+            finally
+            {
+                reader.Depth--;
             }
         }
     }
 
     public class InterfaceReadOnlyReactivePropertyFormatter<T> : IMessagePackFormatter<IReadOnlyReactiveProperty<T>>
     {
-        public int Serialize(ref byte[] bytes, int offset, IReadOnlyReactiveProperty<T> value, IFormatterResolver formatterResolver)
+        public void Serialize(ref MessagePackWriter writer, IReadOnlyReactiveProperty<T> value, MessagePackSerializerOptions options)
         {
             var rxProp = value as ReactiveProperty<T>;
             if (rxProp != null)
             {
-                return ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactiveProperty<T>>().Serialize(ref bytes, offset, rxProp, formatterResolver);
+                ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactiveProperty<T>>().Serialize(ref writer, rxProp, options);
+                return;
             }
 
             var slimProp = value as ReactivePropertySlim<T>;
             if (slimProp != null)
             {
-                return ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactivePropertySlim<T>>().Serialize(ref bytes, offset, slimProp, formatterResolver);
+                ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactivePropertySlim<T>>().Serialize(ref writer, slimProp, options);
+                return;
             }
 
             if (value == null)
             {
-                return MessagePackBinary.WriteNil(ref bytes, offset);
+                writer.WriteNil();
             }
             else
             {
@@ -225,30 +242,38 @@ namespace MessagePack.ReactivePropertyExtension
             }
         }
 
-        public IReadOnlyReactiveProperty<T> Deserialize(byte[] bytes, int offset, IFormatterResolver formatterResolver, out int readSize)
+        public IReadOnlyReactiveProperty<T> Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
         {
-            var length = MessagePackBinary.ReadArrayHeader(bytes, offset, out readSize);
+            var length = reader.ReadArrayHeader();
 
-            switch (length)
+            options.Security.DepthStep(ref reader);
+            try
             {
-                case 2:
-                    return ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactivePropertySlim<T>>().Deserialize(bytes, offset, formatterResolver, out readSize);
-                case 3:
-                    return ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactiveProperty<T>>().Deserialize(bytes, offset, formatterResolver, out readSize);
-                default:
-                    throw new InvalidOperationException("Invalid ReactiveProperty or ReactivePropertySlim data.");
+                switch (length)
+                {
+                    case 2:
+                        return ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactivePropertySlim<T>>().Deserialize(ref reader, options);
+                    case 3:
+                        return ReactivePropertyResolver.Instance.GetFormatterWithVerify<ReactiveProperty<T>>().Deserialize(ref reader, options);
+                    default:
+                        throw new InvalidOperationException("Invalid ReactiveProperty or ReactivePropertySlim data.");
+                }
+            }
+            finally
+            {
+                reader.Depth--;
             }
         }
     }
 
     public class ReactiveCollectionFormatter<T> : CollectionFormatterBase<T, ReactiveCollection<T>>
     {
-        protected override void Add(ReactiveCollection<T> collection, int index, T value)
+        protected override void Add(ReactiveCollection<T> collection, int index, T value, MessagePackSerializerOptions options)
         {
             collection.Add(value);
         }
 
-        protected override ReactiveCollection<T> Create(int count)
+        protected override ReactiveCollection<T> Create(int count, MessagePackSerializerOptions options)
         {
             return new ReactiveCollection<T>();
         }
@@ -256,56 +281,52 @@ namespace MessagePack.ReactivePropertyExtension
 
     public class UnitFormatter : IMessagePackFormatter<Unit>
     {
-        public static IMessagePackFormatter<Unit> Instance = new UnitFormatter();
+        public static readonly UnitFormatter Instance = new UnitFormatter();
 
-        UnitFormatter()
+        private UnitFormatter()
         {
-
         }
 
-        public int Serialize(ref byte[] bytes, int offset, Unit value, IFormatterResolver formatterResolver)
+        public void Serialize(ref MessagePackWriter writer, Unit value, MessagePackSerializerOptions options)
         {
-            return MessagePackBinary.WriteNil(ref bytes, offset);
+            writer.WriteNil();
         }
 
-        public Unit Deserialize(byte[] bytes, int offset, IFormatterResolver formatterResolver, out int readSize)
+        public Unit Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
         {
-            if (MessagePackBinary.IsNil(bytes, offset))
+            if (reader.TryReadNil())
             {
-                readSize = 1;
                 return Unit.Default;
             }
             else
             {
-                throw new InvalidOperationException("Invalid Data type. Code:" + MessagePackCode.ToFormatName(bytes[offset]));
+                throw new InvalidOperationException("Invalid Data type. Code: " + MessagePackCode.ToFormatName(reader.NextCode));
             }
         }
     }
 
     public class NullableUnitFormatter : IMessagePackFormatter<Unit?>
     {
-        public static IMessagePackFormatter<Unit?> Instance = new NullableUnitFormatter();
+        public static readonly NullableUnitFormatter Instance = new NullableUnitFormatter();
 
-        NullableUnitFormatter()
+        private NullableUnitFormatter()
         {
-
         }
 
-        public int Serialize(ref byte[] bytes, int offset, Unit? value, IFormatterResolver formatterResolver)
+        public void Serialize(ref MessagePackWriter writer, Unit? value, MessagePackSerializerOptions options)
         {
-            return MessagePackBinary.WriteNil(ref bytes, offset);
+            writer.WriteNil();
         }
 
-        public Unit? Deserialize(byte[] bytes, int offset, IFormatterResolver formatterResolver, out int readSize)
+        public Unit? Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
         {
-            if (MessagePackBinary.IsNil(bytes, offset))
+            if (reader.TryReadNil())
             {
-                readSize = 1;
                 return Unit.Default;
             }
             else
             {
-                throw new InvalidOperationException("Invalid Data type. Code:" + MessagePackCode.ToFormatName(bytes[offset]));
+                throw new InvalidOperationException("Invalid Data type. Code: " + MessagePackCode.ToFormatName(reader.NextCode));
             }
         }
     }
@@ -313,51 +334,49 @@ namespace MessagePack.ReactivePropertyExtension
     // [Mode, Value]
     public class ReactivePropertySlimFormatter<T> : IMessagePackFormatter<ReactivePropertySlim<T>>
     {
-        public int Serialize(ref byte[] bytes, int offset, ReactivePropertySlim<T> value, IFormatterResolver formatterResolver)
+        public void Serialize(ref MessagePackWriter writer, ReactivePropertySlim<T> value, MessagePackSerializerOptions options)
         {
             if (value == null)
             {
-                return MessagePackBinary.WriteNil(ref bytes, offset);
+                writer.WriteNil();
             }
             else
             {
-                var startOffset = offset;
+                writer.WriteArrayHeader(2);
 
-                offset += MessagePackBinary.WriteFixedArrayHeaderUnsafe(ref bytes, offset, 2);
-
-                offset += MessagePackBinary.WriteInt32(ref bytes, offset, ReactivePropertySchedulerMapper.ToReactivePropertySlimModeInt(value));
-                offset += formatterResolver.GetFormatterWithVerify<T>().Serialize(ref bytes, offset, value.Value, formatterResolver);
-
-                return offset - startOffset;
+                writer.Write(ReactivePropertySchedulerMapper.ToReactivePropertySlimModeInt(value));
+                options.Resolver.GetFormatterWithVerify<T>().Serialize(ref writer, value.Value, options);
             }
         }
 
-        public ReactivePropertySlim<T> Deserialize(byte[] bytes, int offset, IFormatterResolver formatterResolver, out int readSize)
+        public ReactivePropertySlim<T> Deserialize(ref MessagePackReader reader, MessagePackSerializerOptions options)
         {
-            if (MessagePackBinary.IsNil(bytes, offset))
+            if (reader.TryReadNil())
             {
-                readSize = 1;
                 return null;
             }
             else
             {
-                var startOffset = offset;
+                var length = reader.ReadArrayHeader();
+                if (length != 2)
+                {
+                    throw new InvalidOperationException("Invalid ReactivePropertySlim data.");
+                }
 
-                var length = MessagePackBinary.ReadArrayHeader(bytes, offset, out readSize);
-                offset += readSize;
-                if (length != 2) throw new InvalidOperationException("Invalid ReactivePropertySlim data.");
+                options.Security.DepthStep(ref reader);
+                try
+                {
+                    var mode = (ReactivePropertyMode)reader.ReadInt32();
 
-                var mode = (ReactivePropertyMode)MessagePackBinary.ReadInt32(bytes, offset, out readSize);
-                offset += readSize;
+                    T v = options.Resolver.GetFormatterWithVerify<T>().Deserialize(ref reader, options);
 
-                var v = formatterResolver.GetFormatterWithVerify<T>().Deserialize(bytes, offset, formatterResolver, out readSize);
-                offset += readSize;
-
-                readSize = offset - startOffset;
-
-                return new ReactivePropertySlim<T>(v, mode);
+                    return new ReactivePropertySlim<T>(v, mode);
+                }
+                finally
+                {
+                    reader.Depth--;
+                }
             }
-
         }
     }
 }
