@@ -1,4 +1,13 @@
-﻿using MessagePack;
+﻿// Copyright (c) All contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using MessagePack;
+
+#pragma warning disable SA1307 // Field should begin with upper-case letter
+#pragma warning disable SA1300 // Field should begin with upper-case letter
+#pragma warning disable IDE1006 // Field should begin with upper-case letter
+#pragma warning disable SA1649 // type name matches file name
+#pragma warning disable SA1401 // Fields should be private (we need fields rather than auto-properties for .NET Native compilation to work).
 
 namespace UnityEngine
 {
@@ -101,7 +110,6 @@ namespace UnityEngine
         public Color(float r, float g, float b)
             : this(r, g, b, 1.0f)
         {
-
         }
 
         [SerializationConstructor]
@@ -118,10 +126,10 @@ namespace UnityEngine
     public struct Bounds
     {
         [Key(0)]
-        public Vector3 center { get; set; }
+        public Vector3 center;
 
         [IgnoreMember]
-        public Vector3 extents { get; set; }
+        public Vector3 extents;
 
         [Key(1)]
         public Vector3 size
@@ -130,6 +138,7 @@ namespace UnityEngine
             {
                 return this.extents * 2f;
             }
+
             set
             {
                 this.extents = value * 0.5f;
@@ -148,16 +157,16 @@ namespace UnityEngine
     public struct Rect
     {
         [Key(0)]
-        public float x { get; set; }
+        public float x;
 
         [Key(1)]
-        public float y { get; set; }
+        public float y;
 
         [Key(2)]
-        public float width { get; set; }
+        public float width;
 
         [Key(3)]
-        public float height { get; set; }
+        public float height;
 
         [SerializationConstructor]
         public Rect(float x, float y, float width, float height)
@@ -186,98 +195,55 @@ namespace UnityEngine
     }
 
     // additional from 1.7.3.3
-
     [MessagePackObject]
     public sealed class AnimationCurve
     {
         [Key(0)]
-        public Keyframe[] keys { get; set; }
+        public Keyframe[] keys;
 
         [IgnoreMember]
-        public int length { get { return keys.Length; } }
+        public int length
+        {
+            get { return this.keys.Length; }
+        }
 
         [Key(1)]
-        public WrapMode postWrapMode { get; set; }
+        public WrapMode postWrapMode;
 
         [Key(2)]
-        public WrapMode preWrapMode { get; set; }
+        public WrapMode preWrapMode;
     }
 
     [MessagePackObject]
     public struct Keyframe
     {
-        private float m_Time;
-        private float m_Value;
-        private float m_InTangent;
-        private float m_OutTangent;
-
-
         [Key(0)]
-        public float time
-        {
-            get
-            {
-                return this.m_Time;
-            }
-            set
-            {
-                this.m_Time = value;
-            }
-        }
+        public float time;
+
         [Key(1)]
-        public float value
-        {
-            get
-            {
-                return this.m_Value;
-            }
-            set
-            {
-                this.m_Value = value;
-            }
-        }
+        public float value;
 
         [Key(2)]
-        public float inTangent
-        {
-            get
-            {
-                return this.m_InTangent;
-            }
-            set
-            {
-                this.m_InTangent = value;
-            }
-        }
+        public float inTangent;
 
         [Key(3)]
-        public float outTangent
-        {
-            get
-            {
-                return this.m_OutTangent;
-            }
-            set
-            {
-                this.m_OutTangent = value;
-            }
-        }
+        public float outTangent;
 
         public Keyframe(float time, float value)
         {
-            this.m_Time = time;
-            this.m_Value = value;
-            this.m_InTangent = 0f;
-            this.m_OutTangent = 0f;
+            this.time = time;
+            this.value = value;
+            this.inTangent = 0f;
+            this.outTangent = 0f;
         }
 
         [SerializationConstructor]
         public Keyframe(float time, float value, float inTangent, float outTangent)
         {
-            this.m_Time = time;
-            this.m_Value = value;
-            this.m_InTangent = inTangent;
-            this.m_OutTangent = outTangent;
+            this.time = time;
+            this.value = value;
+            this.inTangent = inTangent;
+            this.outTangent = outTangent;
         }
     }
 
@@ -288,7 +254,7 @@ namespace UnityEngine
         PingPong = 4,
         Default = 0,
         ClampForever = 8,
-        Clamp = 1
+        Clamp = 1,
     }
 
     [MessagePackObject]
@@ -332,11 +298,13 @@ namespace UnityEngine
     public sealed class Gradient
     {
         [Key(0)]
-        public GradientColorKey[] colorKeys { get; set; }
+        public GradientColorKey[] colorKeys;
+
         [Key(1)]
-        public GradientAlphaKey[] alphaKeys { get; set; }
+        public GradientAlphaKey[] alphaKeys;
+
         [Key(2)]
-        public GradientMode mode { get; set; }
+        public GradientMode mode;
     }
 
     [MessagePackObject]
@@ -372,7 +340,7 @@ namespace UnityEngine
     public enum GradientMode
     {
         Blend,
-        Fixed
+        Fixed,
     }
 
     [MessagePackObject]
@@ -400,20 +368,19 @@ namespace UnityEngine
     public sealed class RectOffset
     {
         [Key(0)]
-        public int left { get; set; }
+        public int left;
 
         [Key(1)]
-        public int right { get; set; }
+        public int right;
 
         [Key(2)]
-        public int top { get; set; }
+        public int top;
 
         [Key(3)]
-        public int bottom { get; set; }
+        public int bottom;
 
         public RectOffset()
         {
-
         }
 
         public RectOffset(int left, int right, int top, int bottom)
@@ -429,11 +396,10 @@ namespace UnityEngine
     public struct LayerMask
     {
         [Key(0)]
-        public int value { get; set; }
+        public int value;
     }
 
     // from Unity2017.2
-
     [MessagePackObject]
     public struct Vector2Int
     {
@@ -493,16 +459,16 @@ namespace UnityEngine
     public struct RectInt
     {
         [Key(0)]
-        public int x { get; set; }
+        public int x;
 
         [Key(1)]
-        public int y { get; set; }
+        public int y;
 
         [Key(2)]
-        public int width { get; set; }
+        public int width;
 
         [Key(3)]
-        public int height { get; set; }
+        public int height;
 
         [SerializationConstructor]
         public RectInt(int x, int y, int width, int height)
@@ -534,10 +500,10 @@ namespace UnityEngine
     public struct BoundsInt
     {
         [Key(0)]
-        public Vector3Int position { get; set; }
+        public Vector3Int position;
 
         [Key(1)]
-        public Vector3Int size { get; set; }
+        public Vector3Int size;
 
         [SerializationConstructor]
         public BoundsInt(Vector3Int position, Vector3Int size)
