@@ -1,53 +1,46 @@
-﻿using MessagePack.ImmutableCollection;
-using System;
+﻿// Copyright (c) All contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using MessagePack.Formatters;
-using MessagePack.Resolvers;
 using System.Collections.Immutable;
 using Xunit;
 
 namespace MessagePack.Tests.ExtensionTests
 {
-    public class WithImmutableDefaultResolver : IFormatterResolver
-    {
-        public IMessagePackFormatter<T> GetFormatter<T>()
-        {
-            return (ImmutableCollectionResolver.Instance.GetFormatter<T>()
-                 ?? StandardResolver.Instance.GetFormatter<T>());
-        }
-    }
-
     public class ImmutableCollectionTest
     {
-        T Convert<T>(T value)
+        private T Convert<T>(T value)
         {
-            var resolver = new WithImmutableDefaultResolver();
-            return MessagePackSerializer.Deserialize<T>(MessagePackSerializer.Serialize(value, resolver), resolver);
+            MessagePackSerializerOptions options = MessagePackSerializerOptions.Standard.WithResolver(new WithImmutableDefaultResolver());
+            return MessagePackSerializer.Deserialize<T>(MessagePackSerializer.Serialize(value, options), options);
         }
 
-        public static object collectionTestData = new object[]
+        public static object[][] CollectionTestData = new object[][]
         {
-            new object[]{ ImmutableList<int>.Empty.AddRange(new[] { 1, 10, 100 }) , null },
-            new object[]{ ImmutableDictionary<int,int>.Empty.AddRange(new Dictionary<int,int> { { 1, 10 },{ 2, 10 }, { 3, 100 } }) , null },
-            new object[]{ ImmutableHashSet<int>.Empty.Add(1).Add(10).Add(100) , null },
-            new object[]{ ImmutableSortedDictionary<int,int>.Empty.AddRange(new Dictionary<int,int> { { 1, 10 },{ 2, 10 }, { 3, 100 } }) , null },
-            new object[]{ ImmutableSortedSet<int>.Empty.Add(1).Add(10).Add(100) , null },
-            new object[]{ ImmutableQueue<int>.Empty.Enqueue(1).Enqueue(10).Enqueue(100) , null },
-            new object[]{ ImmutableStack<int>.Empty.Push(1).Push(10).Push(100) , null },
+            new object[] { true, ImmutableList<int>.Empty.AddRange(new[] { 1, 10, 100 }), null },
+            new object[] { false, ImmutableDictionary<int, int>.Empty.AddRange(new Dictionary<int, int> { { 1, 10 }, { 2, 10 }, { 3, 100 } }), null },
+            new object[] { false, ImmutableHashSet<int>.Empty.Add(1).Add(10).Add(100), null },
+            new object[] { true, ImmutableSortedDictionary<int, int>.Empty.AddRange(new Dictionary<int, int> { { 1, 10 }, { 2, 10 }, { 3, 100 } }), null },
+            new object[] { true, ImmutableSortedSet<int>.Empty.Add(1).Add(10).Add(100), null },
+            new object[] { true, ImmutableQueue<int>.Empty.Enqueue(1).Enqueue(10).Enqueue(100), null },
+            new object[] { true, ImmutableStack<int>.Empty.Push(1).Push(10).Push(100), null },
         };
 
         [Theory]
-        [MemberData(nameof(collectionTestData))]
-        public void ConcreteCollectionTest<T>(T x, T y)
+        [MemberData(nameof(CollectionTestData))]
+        public void ConcreteCollectionTest<T>(bool ordered, T x, T y)
         {
-            Convert(x).IsStructuralEqual(x);
-            Convert(y).IsStructuralEqual(y);
+            if (ordered)
+            {
+                this.Convert(x).IsStructuralEqual(x);
+            }
+            else
+            {
+                this.Convert(x).IsStructuralEqualIgnoreCollectionOrder(x);
+            }
+
+            this.Convert(y).IsStructuralEqual(y);
         }
-
-
 
         [Fact]
         public void InterfaceCollectionTest()
@@ -58,22 +51,22 @@ namespace MessagePack.Tests.ExtensionTests
             IImmutableQueue<int> d = ImmutableQueue<int>.Empty.Enqueue(1).Enqueue(10).Enqueue(100);
             IImmutableStack<int> e = ImmutableStack<int>.Empty.Push(1).Push(10).Push(100);
 
-            Convert(a).IsStructuralEqual(a);
-            Convert(b).IsStructuralEqual(b);
-            Convert(c).IsStructuralEqual(c);
-            Convert(d).IsStructuralEqual(d);
-            Convert(e).IsStructuralEqual(e);
+            this.Convert(a).IsStructuralEqual(a);
+            this.Convert(b).IsStructuralEqualIgnoreCollectionOrder(b);
+            this.Convert(c).IsStructuralEqualIgnoreCollectionOrder(c);
+            this.Convert(d).IsStructuralEqual(d);
+            this.Convert(e).IsStructuralEqual(e);
 
             a = null;
             b = null;
             c = null;
             d = null;
             e = null;
-            Convert(a).IsNull();
-            Convert(b).IsNull();
-            Convert(c).IsNull();
-            Convert(d).IsNull();
-            Convert(e).IsNull();
+            this.Convert(a).IsNull();
+            this.Convert(b).IsNull();
+            this.Convert(c).IsNull();
+            this.Convert(d).IsNull();
+            this.Convert(e).IsNull();
         }
 
         [Fact]
@@ -83,9 +76,9 @@ namespace MessagePack.Tests.ExtensionTests
             ImmutableArray<int>? b = ImmutableArray.CreateRange(new[] { 1, 10, 100 });
             ImmutableArray<int>? c = null;
 
-            Convert(a).Is(1, 10, 100);
-            Convert(b).Is(1, 10, 100);
-            Convert(c).IsNull();
+            this.Convert(a).Is(1, 10, 100);
+            this.Convert(b).Is(1, 10, 100);
+            this.Convert(c).IsNull();
         }
     }
 }
